pyproject.toml
[tool.poetry]
name = "django-member-system"
version = "0.1.0"
description = ""
authors = ["azzip <azzip-rnd@skvip.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
python-decouple = "^3.8"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
mysqlclient = "^2.2.4"
django-redis = "^5.4.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
